Extract event loop into own class & add handlers
In this way the event loop can be reused for other event streams, as the
message handler can be replaced.
